Fix no traces getting sent on macOS
Yes, this change is absolutely baffling why it fixes the bug, but it does. I
think that what happened is that even though both
/run/current-system/sw/bin/nix and the nix used by the flake are 2.11, we are
still doing some kind of ABI crimes.

It seems that some kind of macOS shenanigans means that our copy of the
nix::logger symbol is different than the one that Nix is using because of this.

The solution here is to *ensure* that the nix you call the plugin with is the
same nix as it was built with, which we probably should have been doing in the
first place anyway!